refactor: separate ChatHistory and ChatInput components
Fixes #227

Splits ChatInterface into two components with distinct lifecycles:

- ChatHistory: Contains ChatQueue (with ink's Static component).
  MUST always be rendered to prevent memory leaks when modal modes activate.

- ChatInput: Contains user input, tool confirmation, execution indicators.
  Can safely mount/unmount when modal dialogs are shown.

This architectural change prevents the memory leak that occurred when modal
modes (config wizard, model selector, etc.) would unmount ChatInterface,
causing ink's Static component to be destroyed and recreated.

Also renames component files to kebab-case for consistency:
- AppContainer → app-container
- ModalSelectors → modal-selectors
- ChatInterface → deleted (replaced by chat-history + chat-input)
